refactor: merge delete
