--- internal/handler/leasing_history_handler_delete.go
@@ -1,38 +1,0 @@
-package handler
-
-import (
-	"github.com/PitiNarak/condormhub-backend/pkg/apperror"
-	"github.com/gofiber/fiber/v2"
-	"github.com/google/uuid"
-)
-
-// Delete godoc
-// @Summary Delete a leasing history
-// @Description Delete a leasing history in the database
-// @Tags history
-// @Security Bearer
-// @Produce json
-// @Param id path string true "LeasingHistoryId"
-// @Success 204 "No Content"
-// @Failure 400 {object} dto.ErrorResponse "Incorrect UUID format"
-// @Failure 401 {object} dto.ErrorResponse "your request is unauthorized"
-// @Failure 404 {object} dto.ErrorResponse "leasing history not found"
-// @Failure 500 {object} dto.ErrorResponse "Can not parse UUID or Failed to delete leasing history"
-// @Router /history/{id} [delete]
-func (h *LeasingHistoryHandler) Delete(c *fiber.Ctx) error {
-	id := c.Params("id")
-
-	if err := uuid.Validate(id); err != nil {
-		return apperror.BadRequestError(err, "Incorrect UUID format")
-	}
-
-	leasingHistoryID, err := uuid.Parse(id)
-	if err != nil {
-		return apperror.InternalServerError(err, "Can not parse UUID")
-	}
-	err = h.service.Delete(leasingHistoryID)
-	if err != nil {
-		return err
-	}
-	return c.SendStatus(fiber.StatusNoContent)
-}
